requirements.txt
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
autoflake==2.3.0
Automat==22.10.0
black==24.2.0
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
channels==4.0.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
cryptography==42.0.4
daphne==4.1.0
distlib==0.3.8
Django==5.0.2
django-oauth-toolkit==2.3.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
filelock==3.13.1
hyperlink==21.0.0
identify==2.5.35
idna==3.6
incremental==22.10.0
isort==5.13.2
jwcrypto==1.5.4
mypy-extensions==1.0.0
nodeenv==1.8.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
pre-commit==3.6.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyflakes==3.2.0
pyOpenSSL==24.0.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.9.0
urllib3==2.2.1
virtualenv==20.25.1
zope.interface==6.2
